requirements.txt
alembic==0.9.7
attrs==17.4.0
cffi==1.11.4
click==6.7
cryptography==1.7.2
dominate==2.3.1
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-Moment==0.6.0
Flask-SendGrid==0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
munkres==1.0.12
numpy==1.14.0
pandas==0.22.0
pg8000==1.11.0
pluggy==0.6.0
psycopg2==2.7.3.2
py==1.5.2
pyasn1==0.4.2
pycparser==2.18
PyJWT==1.5.3
pyScss==1.3.5
pytest==3.3.2
python-dateutil==2.6.1
python-editor==1.0.3
python-http-client==3.0.0
pytz==2017.3
PyYAML==3.12
sendgrid==5.3.0
six==1.11.0
SQLAlchemy==1.2.1
testing.common.database==2.0.3
testing.postgresql==1.3.0
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.1
